fix(lti): move rateLimit module out of api/_lib/ to fix Node.js cold-start
